Lesson 1.2 TodoList Assignment
